pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hdt-agentic-interop"
version = "0.5.0"
description = "Interoperable HDT prototype with MCP façade, vault, policy lanes, and telemetry."
readme = "README.md"
requires-python = ">=3.11,<3.15"
license = { text = "MIT" }
authors = [{ name = "Olga Glazunova", email = "o.glazunova@tue.nl" }]
keywords = ["Model Context Protocol", "HDT", "interoperability", "digital health", "agentic systems"]

dependencies = [
  "flask>=3.0",
  "requests>=2.31",
  "python-dotenv>=1.0",
  "mcp>=1.0.0",
  # FastMCP is used by the MCP server implementation
  "fastmcp==2.13.0.2",
  "pydantic>=2.5",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.0",
  "pre-commit>=3.6",
  "ruff>=0.6.0",
  "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/oaglazunova/HDT-agentic-interop"
Repository = "https://github.com/oaglazunova/HDT-agentic-interop"
Issues = "https://github.com/oaglazunova/HDT-agentic-interop/issues"

# Optional convenience script (adjust the module path if your server file differs)
[project.scripts]
hdt-mcp = "hdt_mcp.server:main"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["hdt_mcp", "hdt_api", "tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
strict_optional = false
